refactor: Renombrado de clases y reorganización de estructura de productos
- Renombrado 'organicProduct' a 'OrganicProduct' (PascalCase). - Corregida la llamada al constructor de 'Vegetables' en OrganicProductServices. - Eliminadas clases de producto y proveedor antiguas/duplicadas. - Reorganizada la estructura de archivos en 'model/Product' y 'model/Supplier'.
