+++ src/services/OrganicProductServices.java
@@ -1,5 +1,31 @@
 package services;
 
-public class OrganicProductServices {
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
+import model.Product.OrganicProduct;
+import model.Product.Vegetables;
+
+public class OrganicProductServices {
+    private List<OrganicProduct> _organicProduct;
+    private int nextId;
+
+    public OrganicProductServices() {
+        this._organicProduct = new ArrayList<>();
+        this.nextId = 1;
+    }
+
+    // ?Metodo para crear un producto de tipo vegetal
+    public void createVegetableProduct(String nameProduct, String descriptionProduct, int idSupplier,
+            LocalDateTime dateAdmission, String type, double price, String vegetableType, int freshnessDays) {
+        Vegetables newVegetables = new Vegetables(nextId++, nameProduct, descriptionProduct, idSupplier, dateAdmission,
+                type, price, vegetableType, freshnessDays);
+        _organicProduct.add(newVegetables);
+    }
+
+    // ?Metodo para obtener todo los productos
+    public List<OrganicProduct> getAllOrganicProducts() {
+        return new ArrayList<>(_organicProduct);
+    }
 }
